rename confirm dialog files and folder
