+++ src/app/components/shared-dialogs/text-input-dialog/text-input-dialog-options.ts
@@ -1,11 +1,7 @@
 export interface TextInputDialogOptions {
     title: string;
-
     message?: string;
-
     label?: string;
-
     format?: string | RegExp;
-
     value?: string;
 }
