+++ src/app/components/shared-dialogs/text-input-dialog/text-input-dialog.component.ts
@@ -1,7 +1,7 @@
 import { Component, Inject } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
-import { TextInputDialogOptions } from './TextInputDialogOptions';
+import { TextInputDialogOptions } from './text-input-dialog-options';
 
 interface InputTextDialogModel {
     value: string;
